django_demo/demosite_docker3/demosite/requirements.txt
asgiref==3.3.4
cachetools==4.2.2
certifi==2021.10.8
cffi==1.14.6
charset-normalizer==2.0.6
cryptography==3.4.7
cycler==0.10.0
dj-database-url==0.5.0
dj-static==0.0.6
Django==3.2.4
django-appconf==1.0.4
django-debug-toolbar==3.2.2
django-filter==21.1
django-googledrive-storage==1.5.2
django-heroku==0.3.1
djangorestframework==3.12.4
fonttools==4.25.0
google-api-core==2.0.1
google-api-python-client==2.22.0
google-auth==2.1.0
google-auth-httplib2==0.1.0
googleapis-common-protos==1.53.0
gunicorn==20.1.0
httplib2==0.19.1
idna==3.2
importlib-metadata==4.8.2
Jinja2==3.0.3
kiwisolver==1.3.1
Markdown==3.3.6
MarkupSafe==2.0.1
matplotlib==3.4.3
micropython-fcntl==0.0.3
micropython-ffilib==0.1.2
mkl-fft==1.3.1
mkl-random==1.2.2
mkl-service==2.4.0
munkres==1.1.4
Naked==0.1.31
numpy==1.21.2
oauth2client==4.1.3
olefile==0.46
Pillow==8.4.0
prettytable==2.4.0
protobuf==3.18.0
psycopg2==2.8.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pycryptodome==3.11.0
pyecharts==1.9.1
PyMySQL==1.0.2
pyparsing==2.4.7
python-dateutil==2.8.2
pytz==2021.1
PyYAML==6.0
reportlab==3.6.2
requests==2.26.0
rsa==4.7.2
shellescape==3.8.1
simplejson==3.17.6
sip==5.0.0
six==1.16.0
sqlparse==0.4.1
static3==0.7.0
tornado==6.1
uritemplate==3.0.1
urllib3==1.26.7
wcwidth==0.2.5
whitenoise==5.3.0
wincertstore==0.2
zipp==3.6.0
uwsgi
